feat(즐겨찾기) : 즐겨찾기(mark) 리패키징에 따른 경로 수정
